requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-crispy-forms==1.9.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
mistune==0.8.4
Pygments==2.6.1
python-dateutil==1.5
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
